requirements.txt
Flask
roboflow
opencv-python-headless
google-cloud-vision
numpy
Pillow
